refactor: modularize telemetry into separate concerns
Refactors the `telemetry.ts` module into multiple, more focused files within the `packages/core/src/telemetry/` directory. This change improves organization and maintainability by separating concerns such as type definitions, constants, metrics, logging, and SDK management.
